# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionItemCreate": {
        "additionalProperties": false,
        "properties": {
          "assignee_participant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Participant Id"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "start_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ms"
          },
          "text": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ActionItemCreate",
        "type": "object"
      },
      "ActionItemRead": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssigneeRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "meeting_id": {
            "title": "Meeting Id",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/ActionItemSource"
          },
          "start_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ms"
          },
          "status": {
            "$ref": "#/components/schemas/ActionItemStatus"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "meeting_id",
          "text",
          "status",
          "assignee",
          "due_date",
          "completed_at",
          "source",
          "start_ms"
        ],
        "title": "ActionItemRead",
        "type": "object"
      },
      "ActionItemSource": {
        "enum": [
          "ai",
          "manual"
        ],
        "title": "ActionItemSource",
        "type": "string"
      },
      "ActionItemStatus": {
        "enum": [
          "open",
          "completed"
        ],
        "title": "ActionItemStatus",
        "type": "string"
      },
      "ActionItemUpdate": {
        "additionalProperties": false,
        "description": "Partial edit; services use model_fields_set so null can clear assignee or due date.",
        "properties": {
          "assignee_participant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Participant Id"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionItemStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "title": "ActionItemUpdate",
        "type": "object"
      },
      "AssigneeRead": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "AssigneeRead",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Page_ActionItemRead_": {
        "properties": {
          "has_next": {
            "readOnly": true,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "readOnly": true,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next"
        ],
        "title": "Page[ActionItemRead]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fireflies Clone API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/action-items/{item_id}": {
      "delete": {
        "operationId": "delete_action_item",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Delete an action item",
        "tags": [
          "action-items"
        ]
      },
      "patch": {
        "operationId": "update_action_item",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Update an action item",
        "tags": [
          "action-items"
        ]
      }
    },
...
    "/api/v1/meetings/{meeting_id}/action-items": {
      "get": {
        "operationId": "list_action_items",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Clamped to 100.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Clamped to 100.",
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ActionItemRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "List a meeting's action items",
        "tags": [
          "action-items"
        ]
      },
      "post": {
        "operationId": "create_action_item",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Add an action item",
        "tags": [
          "action-items"
        ]
      }
    },
...
  }
}